refactor(viewer): extract FEA warp, beam-solid and element-edge sidecars from the loader
Four more fea/streaming modules: sceneMesh (find the GLB's mesh,
snapshot base positions, install the AFEM selection userData), warp
(resolve the morph source, lerp the beam-solid warp, share morph state
with wireframe children), beamSolids (fetch + upgrade the beam-solid
GLB, fetch its AFBV mapping) and elementEdges (the shell and beam edge
overlays). The two inline blocks the loader carried for AFBV and edges
become functions that return what they fetched; writing it onto the
session stays in the loader. No behaviour change.
