+++ src/frontend/src/utils/scene/fea/streaming/sceneMesh.ts
@@ -1,0 +1,101 @@
+// FEA streaming: the result mesh as a scene object.
+//
+// Owns: finding the mesh inside a loaded GLB, snapshotting its base positions,
+// and the userData the worker cache reads to wire per-element selection
+// (draw ranges + id hierarchy from the AFEM sidecar). Pure functions over the
+// objects handed in; nothing here touches a store or the session.
+//
+// Inputs: a THREE object tree and the parsed AFEM entries.
+
+import * as THREE from "three";
+
+import type {MeshElementEntry} from "@/services/feaMeshElements";
+
+export function findFirstMesh(root: THREE.Object3D): THREE.Mesh | null {
+    let found: THREE.Mesh | null = null;
+    root.traverse((obj) => {
+        if (found) return;
+        if ((obj as THREE.Mesh).isMesh) {
+            found = obj as THREE.Mesh;
+        }
+    });
+    return found;
+}
+
+export function snapshotBasePositions(geometry: THREE.BufferGeometry): Float32Array {
+    const attr = geometry.getAttribute("position");
+    if (!attr || attr.itemSize !== 3) {
+        throw new Error("FEA mesh GLB has no usable position attribute");
+    }
+    return new Float32Array(attr.array as Float32Array);
+}
+
+/** Build the draw-ranges + id-hierarchy userData entries that
+ * prepareLoadedModel reads to wire CustomBatchedMesh selection.
+ *
+ * AFEM stores triangles per element; the index buffer counts vertex
+ * indices, so we multiply ``tri_start`` and ``tri_count`` by 3 here.
+ * Line elements (``triCount === 0``) are dropped from the draw-range
+ * map but kept in id_hierarchy so name resolution still works for
+ * them (selection won't fire on them via the triangle picker yet —
+ * Phase 1.A doesn't wire line-element selection).
+ *
+ * The mesh is renamed to ``node0`` because the worker cache filter
+ * (``cacheModelUtils.ts``) only consumes userData keys prefixed
+ * ``draw_ranges_node`` — a quirk of how the CAD GLB pipeline names
+ * primitives (``node0``, ``node0_1``, etc.). Without the rename,
+ * the worker has no draw-range cache → ``queryMeshDrawRange``
+ * returns null → click selection silently does nothing. The
+ * id_hierarchy uses a synthetic root entry ``fea-root`` (parent
+ * ``"*"``, the worker's root sentinel) so every element has a
+ * resolvable parent.
+ */
+export function installAfemUserData(
+    gltf_scene: THREE.Group,
+    entries: MeshElementEntry[],
+): void {
+    const mesh = findFirstMesh(gltf_scene);
+    if (!mesh) {
+        // Could be a line-only mesh exported as a PointCloud — no
+        // selection to wire.
+        return;
+    }
+
+    // Rename to a name the worker filter accepts. The filter is
+    // hard-coded for "draw_ranges_node*" prefixes; renaming here is
+    // less invasive than relaxing the filter.
+    mesh.name = "node0";
+    const finalName = mesh.name;
+
+    // Tag the mesh so prepareLoadedModel skips the design-side edge
+    // overlay (CustomBatchedMesh.getEdgeOverlay). That overlay is
+    // built from originalGeometry with a static applyMatrix4 and
+    // doesn't share morph attribute / influences, so it'd stay at
+    // the un-deformed position while the face mesh + our AFEM-derived
+    // wireframe morph. The AFEM wireframe already shows element
+    // boundaries; per-edge selection highlight via the design-edge
+    // shader isn't wired up for the streaming mesh anyway.
+    mesh.userData.feaStreaming = true;
+
+    const drawRanges: Record<string, [number, number]> = {};
+    const idHierarchy: Record<string, [string, string | number]> = {};
+
+    // Synthetic root: every element points to this as its parent.
+    // The worker's root sentinel is "*" — without a concrete root
+    // node referenced by parent="*", the tree would have multiple
+    // roots and only the last element processed would survive as
+    // the visible tree.
+    const ROOT_RANGE_ID = "fea-root";
+    idHierarchy[ROOT_RANGE_ID] = ["FEA elements", "*"];
+
+    for (const entry of entries) {
+        const rangeId = `E${entry.label}`;
+        idHierarchy[rangeId] = [rangeId, ROOT_RANGE_ID];
+        if (entry.triCount > 0) {
+            drawRanges[rangeId] = [entry.triStart * 3, entry.triCount * 3];
+        }
+    }
+
+    gltf_scene.userData[`draw_ranges_${finalName}`] = drawRanges;
+    gltf_scene.userData["id_hierarchy"] = idHierarchy;
+}
