+++ src/frontend/src/utils/scene/handlers/load_fea_streaming.ts
@@ -1,17 +1,10 @@
 import * as THREE from "three";
 
 import {SceneOperations} from "@/flatbuffers/scene/scene-operations";
-import {GLTFLoader} from "three/examples/jsm/loaders/GLTFLoader";
 
-import {cacheAndBuildTree} from "@/state/model_worker/cacheModelUtils";
 import {fetchElemFieldStep} from "@/services/feaElemFieldBlob";
 import {fetchFieldStep, makeViewerApiFetcher} from "@/services/feaFieldBlob";
-import type {FeaFetcher, FeaRangeFetcher} from "@/services/fea/feaFetcher";
-import {fetchBeamSolidsWarp, ParsedBeamSolidsWarp} from "@/services/feaBeamSolidsWarp";
-import {fetchMeshEdges} from "@/services/feaMeshEdges";
 import {fetchMeshElements, MeshElementEntry} from "@/services/feaMeshElements";
-import {convert_to_custom_batch_mesh} from "@/utils/scene/convert_to_custom_batch_mesh";
-import {clipWithModel} from "@/utils/scene/section_clipping";
 import type {FeaManifest, FeaManifestField} from "@/services/viewerApi";
 import {capabilities} from "@/services/capabilities";
 import {runResultSidecarLoaders} from "@/plugins/sidecarLoaders";
@@ -24,23 +17,23 @@ import {useAnimationStore} from "@/state/animationStore";
 import {useFeaAnimationStore} from "@/state/feaAnimationStore";
 import {useColorStore} from "@/state/colorLegendStore";
 import {useConversionStore} from "@/state/conversionStore";
-import {usePerfStore, requestRender} from "@/state/perfStore";
 import {applyFieldToMesh} from "../fea/applyField";
 import {applyElemFieldToMesh} from "../fea/applyElemField";
 import {resolveContourRange} from "../fea/contourScale";
 import {selectedResultRange} from "../fea/resultUnits";
-import {translationOffsets, warpValue} from "../fea/warpComponents";
 import {autoWarpScale} from "../fea/warpScale";
 import {
     noteFieldSourceLoaded,
     requestingSceneColorOwner,
 } from "../fea/modeSceneColor";
 import {beamSolidNodalColors} from "../fea/beamSolidNodalColors";
-import {FEA_BEAM_EDGE_COLOR, FEA_EDGE_COLOR} from "../fea/edgeColors";
-import {withoutEdges} from "../fea/edgeSplit";
 import {expandSourceTriples, sourceVertexIndices} from "../fea/elementLocalGeometry";
 import {replace_model} from "./update_scene_from_message";
 import {feaSession as session} from "../fea/streaming/session";
+import {findFirstMesh, installAfemUserData, snapshotBasePositions} from "../fea/streaming/sceneMesh";
+import {findDisplacementField, installBeamSolidWarp, linkLineMorphToMesh, resolveWarpSource} from "../fea/streaming/warp";
+import {fetchBeamSolidWarpSidecar, tryLoadBeamSolids} from "../fea/streaming/beamSolids";
+import {installElementEdges} from "../fea/streaming/elementEdges";
 import {refreshUndeformedGhost, setFeaResultColorsVisible, syncFeaOverlayVisibility} from "../fea/streaming/visibility";
 
 export {
@@ -60,400 +53,6 @@ export {
     setFeaUndeformedGhost,
     syncFeaOverlayVisibility,
 } from "../fea/streaming/visibility";
-
-function findFirstMesh(root: THREE.Object3D): THREE.Mesh | null {
-    let found: THREE.Mesh | null = null;
-    root.traverse((obj) => {
-        if (found) return;
-        if ((obj as THREE.Mesh).isMesh) {
-            found = obj as THREE.Mesh;
-        }
-    });
-    return found;
-}
-
-/** Build the draw-ranges + id-hierarchy userData entries that
- * prepareLoadedModel reads to wire CustomBatchedMesh selection.
- *
- * AFEM stores triangles per element; the index buffer counts vertex
- * indices, so we multiply ``tri_start`` and ``tri_count`` by 3 here.
- * Line elements (``triCount === 0``) are dropped from the draw-range
- * map but kept in id_hierarchy so name resolution still works for
- * them (selection won't fire on them via the triangle picker yet —
- * Phase 1.A doesn't wire line-element selection).
- *
- * The mesh is renamed to ``node0`` because the worker cache filter
- * (``cacheModelUtils.ts``) only consumes userData keys prefixed
- * ``draw_ranges_node`` — a quirk of how the CAD GLB pipeline names
- * primitives (``node0``, ``node0_1``, etc.). Without the rename,
- * the worker has no draw-range cache → ``queryMeshDrawRange``
- * returns null → click selection silently does nothing. The
- * id_hierarchy uses a synthetic root entry ``fea-root`` (parent
- * ``"*"``, the worker's root sentinel) so every element has a
- * resolvable parent.
- */
-function installAfemUserData(
-    gltf_scene: THREE.Group,
-    entries: MeshElementEntry[],
-): void {
-    const mesh = findFirstMesh(gltf_scene);
-    if (!mesh) {
-        // Could be a line-only mesh exported as a PointCloud — no
-        // selection to wire.
-        return;
-    }
-
-    // Rename to a name the worker filter accepts. The filter is
-    // hard-coded for "draw_ranges_node*" prefixes; renaming here is
-    // less invasive than relaxing the filter.
-    mesh.name = "node0";
-    const finalName = mesh.name;
-
-    // Tag the mesh so prepareLoadedModel skips the design-side edge
-    // overlay (CustomBatchedMesh.getEdgeOverlay). That overlay is
-    // built from originalGeometry with a static applyMatrix4 and
-    // doesn't share morph attribute / influences, so it'd stay at
-    // the un-deformed position while the face mesh + our AFEM-derived
-    // wireframe morph. The AFEM wireframe already shows element
-    // boundaries; per-edge selection highlight via the design-edge
-    // shader isn't wired up for the streaming mesh anyway.
-    mesh.userData.feaStreaming = true;
-
-    const drawRanges: Record<string, [number, number]> = {};
-    const idHierarchy: Record<string, [string, string | number]> = {};
-
-    // Synthetic root: every element points to this as its parent.
-    // The worker's root sentinel is "*" — without a concrete root
-    // node referenced by parent="*", the tree would have multiple
-    // roots and only the last element processed would survive as
-    // the visible tree.
-    const ROOT_RANGE_ID = "fea-root";
-    idHierarchy[ROOT_RANGE_ID] = ["FEA elements", "*"];
-
-    for (const entry of entries) {
-        const rangeId = `E${entry.label}`;
-        idHierarchy[rangeId] = [rangeId, ROOT_RANGE_ID];
-        if (entry.triCount > 0) {
-            drawRanges[rangeId] = [entry.triStart * 3, entry.triCount * 3];
-        }
-    }
-
-    gltf_scene.userData[`draw_ranges_${finalName}`] = drawRanges;
-    gltf_scene.userData["id_hierarchy"] = idHierarchy;
-}
-
-/** Pick the displacement field from the manifest. Frontend reads
- *  ``category`` set by the bake to find it without re-string-matching
- *  solver-specific names. Returns the first match or null. */
-function findDisplacementField(manifest: FeaManifest): FeaManifestField | null {
-    for (const f of manifest.fields) {
-        if (f.category === "displacement") return f;
-    }
-    return null;
-}
-
-/** Resolve which field (and which step-values) drives the morph
- *  delta for this apply. The colour field is always the user's pick;
- *  warp is decoupled so stress / strain visualisations can still show
- *  the deformed shape. Returns ``null`` when the geometry should stay
- *  static (reaction fields, warp toggle off + no displacement field
- *  available, or the user picked displacement but warpEnabled is off). */
-async function resolveWarpSource(
-    rangeFetcher: FeaRangeFetcher,
-    fetcher: FeaFetcher,
-    cacheKey: string,
-    manifest: FeaManifest,
-    colorField: FeaManifestField,
-    stepIndex: number,
-    warpEnabled: boolean,
-): Promise<{field: FeaManifestField; stepValues: Float32Array} | null> {
-    // Reaction force fields never drive a deformation — applying them
-    // as a morph delta would visualise a force vector as a
-    // displacement, which is semantically wrong. Lock off regardless
-    // of the toggle.
-    if (colorField.category === "reaction") return null;
-
-    // For the displacement field itself, the warp toggle still
-    // controls whether the user sees the deformed shape — a user
-    // inspecting raw DX values may want them on the un-deformed mesh.
-    if (colorField.category === "displacement") {
-        if (!warpEnabled) return null;
-        const stepValues = await fetchFieldStep(rangeFetcher, fetcher, colorField, stepIndex, cacheKey);
-        return {field: colorField, stepValues};
-    }
-
-    // Stress / strain / other — warp by the manifest's displacement
-    // field when the user has the toggle on.
-    if (!warpEnabled) return null;
-    const dispField = findDisplacementField(manifest);
-    if (!dispField) return null;
-
-    // Step alignment: prefer the same index. If the displacement
-    // field has fewer steps (rare — sub-step output), clamp to last.
-    let warpStep = stepIndex;
-    if (warpStep >= dispField.n_steps) {
-        warpStep = dispField.n_steps - 1;
-        // eslint-disable-next-line no-console
-        console.warn(
-            `[fea-streaming] colour-field step ${stepIndex} exceeds displacement-field ` +
-            `n_steps=${dispField.n_steps}; clamping warp source to step ${warpStep}`,
-        );
-    }
-    const stepValues = await fetchFieldStep(rangeFetcher, fetcher, dispField, warpStep, cacheKey);
-    return {field: dispField, stepValues};
-}
-
-/** Fetch + parse the beam-solid GLB and its AFEM sidecar, returning a
- *  THREE.Mesh ready to attach to the scene with per-beam drawRanges
- *  already installed. Returns ``null`` if the manifest carries no
- *  beam-solid URL or the fetch failed (logged + non-fatal). */
-async function tryLoadBeamSolids(
-    fetcher: FeaFetcher,
-    sourceName: string,
-    manifest: FeaManifest,
-    initialVisible: boolean,
-): Promise<{mesh: THREE.Mesh; basePositions: Float32Array} | null> {
-    const beamGlbUrl = manifest.mesh.beam_solids_url;
-    if (!beamGlbUrl) return null;
-    // Perf-store opt-out: when the user wants to A/B against the
-    // line-element fallback we skip the GLB fetch + AFEM/AFBV parsing
-    // entirely. Toggled live via the Performance panel; takes effect
-    // on the next FEA stream load.
-    if (usePerfStore.getState().hideBeamSolids) {
-        return null;
-    }
-
-    try {
-        const [buf, afemEntries] = await Promise.all([
-            fetcher(beamGlbUrl),
-            manifest.mesh.beam_solids_elements_url
-                ? fetchMeshElements(fetcher, manifest.mesh.beam_solids_elements_url)
-                : Promise.resolve<MeshElementEntry[]>([]),
-        ]);
-        const blob = new Blob([buf], {type: "model/gltf-binary"});
-        const url = URL.createObjectURL(blob);
-        let gltfMesh: THREE.Mesh | null = null;
-        try {
-            const loader = new GLTFLoader();
-            const gltf = await new Promise<{scene: THREE.Group}>((resolve, reject) => {
-                loader.load(url, resolve as never, undefined, reject);
-            });
-            gltfMesh = findFirstMesh(gltf.scene);
-        } finally {
-            URL.revokeObjectURL(url);
-        }
-        if (!gltfMesh) return null;
-
-        // Build the draw-range Map keyed by ``E${label}`` so the AFEL
-        // apply kernel and the click-resolver both find ranges with
-        // the same lookup as the main mesh.
-        const drawRanges = new Map<string, [number, number]>();
-        for (const entry of afemEntries) {
-            if (entry.triCount > 0) {
-                drawRanges.set(`E${entry.label}`, [
-                    entry.triStart * 3,
-                    entry.triCount * 3,
-                ]);
-            }
-        }
-        // Rename to ``node1`` so the worker-cache filter accepts the
-        // companion userData key. The main mesh is ``node0`` —
-        // distinct names keep the two meshes' draw-range tables
-        // separate in the worker cache.
-        gltfMesh.name = "node1";
-        gltfMesh.userData.feaBeamSolids = true;
-        gltfMesh.userData.feaStreaming = true;
-        gltfMesh.visible = initialVisible;
-
-        // Upgrade to a CustomBatchedMesh so clicks resolve through
-        // the existing picker pipeline (handleClickMesh → drawRanges
-        // → range_id). Without this, raycasts hit a plain Mesh that
-        // has no ``unique_key`` and the selection silently no-ops.
-        const uniqueKey = `fea-beam-solids::${sourceName}`;
-        const custom = convert_to_custom_batch_mesh(
-            gltfMesh,
-            drawRanges,
-            uniqueKey,
-            /* is_design */ false,
-            /* ada_ext_data */ null,
-        );
-        // Preserve the userData tags + visibility flags the plain
-        // mesh carried; convert_to_custom_batch_mesh copies userData
-        // but it's worth being explicit so future tags don't get
-        // lost to a helper refactor.
-        custom.userData.feaBeamSolids = true;
-        custom.userData.feaStreaming = true;
-        custom.visible = initialVisible;
-
-        // Register with the off-thread worker cache so the picker's
-        // ``queryMeshDrawRange(unique_key, "node1", faceIndex)`` finds
-        // the range and ``queryNameFromRangeId(unique_key, rangeId)``
-        // returns the element label. Synthetic id_hierarchy with a
-        // single FEA-beam root keeps name resolution flat — every
-        // beam shows up as ``E${label}`` in the info box.
-        const hierarchy: Record<string, [string, string | number]> = {};
-        const rangesPlain: Record<string, [number, number]> = {};
-        const ROOT_KEY = "fea-beam-solids-root";
-        hierarchy[ROOT_KEY] = ["Beam solids", "*"];
-        for (const entry of afemEntries) {
-            if (entry.triCount > 0) {
-                const rid = `E${entry.label}`;
-                hierarchy[rid] = [rid, ROOT_KEY];
-                rangesPlain[rid] = [entry.triStart * 3, entry.triCount * 3];
-            }
-        }
-        // The Outliner resolves a clicked row through the runtime ``modelKeyMap``:
-        // ``model_key`` -> an object whose subtree holds the named mesh.
-        // ``setupModelLoader`` registers the main FEA mesh when it loads the GLB;
-        // nothing registered this one. So clicking a beam in the tree set the
-        // Properties name and made NO 3d selection -- the status bar stayed on
-        // "No selection", nothing highlighted, and every selection-driven
-        // behaviour was silently skipped for beams.
-        const keyMapRef = getViewerRuntime().modelKeyMap;
-        if (!keyMapRef.current) keyMapRef.current = new Map();
-        keyMapRef.current.set(uniqueKey, custom);
-
-        // Best-effort cache install — if it fails, the mesh still
-        // renders, the click just won't resolve.
-        void cacheAndBuildTree(uniqueKey, {
-            id_hierarchy: hierarchy,
-            draw_ranges_node1: rangesPlain,
-        });
-
-        // Don't flip vertexColors on here — without a color attribute,
-        // three.js renders vertexColors=true geometry as black. The
-        // AFEL apply kernel turns vertexColors on at the same time it
-        // writes the color attribute, so the first paint lands both
-        // together. Until then the GLB's base PBR material colour
-        // shows, which is the right "no data" state for solid beams.
-
-        const basePositions = snapshotBasePositions(custom.geometry);
-        return {mesh: custom, basePositions};
-    } catch (err) {
-        // Beam-solid rendering is decorative — log and continue so a
-        // missing/corrupt GLB doesn't block rendering of the main mesh.
-        // eslint-disable-next-line no-console
-        console.warn("[fea-streaming] failed to load beam-solid mesh:", err);
-        return null;
-    }
-}
-
-/** Install the beam-solid mesh's morph delta from a nodal
- *  displacement field. Per vertex:
- *
- *    delta_v = lerp(disp[node0], disp[node1], t) × (only first 3 components)
- *
- *  Linked to the main mesh's ``morphTargetInfluences`` so the slider
- *  drives both meshes in lockstep. No-op when the active session
- *  has no beam-solid mesh or no AFBV mapping. */
-function installBeamSolidWarp(
-    main: THREE.Mesh,
-    beamSolid: THREE.Mesh,
-    basePositions: Float32Array,
-    warp: ParsedBeamSolidsWarp,
-    warpField: FeaManifestField | undefined,
-    warpStepValues: Float32Array | undefined,
-): void {
-    const nVerts = warp.n_verts;
-    const displacement = new Float32Array(nVerts * 3);
-
-    if (warpField && warpStepValues) {
-        const nc = warpField.components.length;
-        // WHICH slots hold the translation. A Sesam displacement field is
-        // ["ALL","X","Y","Z","RX","RY","RZ"] -- reading slots 0..2 warps every
-        // vertex by (ALL, X, Y), and since `ALL` is a non-negative aggregate the
-        // beams visibly fly off. See translationOffsets.
-        const axes = translationOffsets(warpField);
-        const n0 = warp.node0;
-        const n1 = warp.node1;
-        const ts = warp.t;
-        for (let v = 0; v < nVerts; v++) {
-            const t = ts[v];
-            const a = n0[v] * nc;
-            const b = n1[v] * nc;
-            const out = v * 3;
-            const ax = warpValue(warpStepValues, a, axes[0]);
-            const ay = warpValue(warpStepValues, a, axes[1]);
-            const az = warpValue(warpStepValues, a, axes[2]);
-            const bx = warpValue(warpStepValues, b, axes[0]);
-            const by = warpValue(warpStepValues, b, axes[1]);
-            const bz = warpValue(warpStepValues, b, axes[2]);
-            const omt = 1 - t;
-            displacement[out + 0] = omt * ax + t * bx;
-            displacement[out + 1] = omt * ay + t * by;
-            displacement[out + 2] = omt * az + t * bz;
-        }
-    }
-    // Else: leave displacement at zero — no warp source means no
-    // deformation, which is what the user gets when they pick a
-    // reaction field or turn warp off.
-
-    const geom = beamSolid.geometry;
-
-    // In the geometry's CURRENT vertex numbering, which is not always the one the
-    // warp sidecar is written in. Painting an element field expands this geometry to
-    // element-local vertices and swaps its position buffer -- 33,812 vertices become
-    // 172,260 -- and the expansion is cached, so it survives a switch back to a nodal
-    // field. A morph sized for the original count is silently ignored by three.js,
-    // which is why beam solids sat undeformed while everything around them moved.
-    // `sourceVertexIndices` returns the cached map, or the identity when the geometry
-    // was never expanded, so this is a no-op on the untouched case.
-    const renderToSource = sourceVertexIndices(geom, nVerts);
-    const renderPositions = expandSourceTriples(basePositions, renderToSource);
-    const renderDisplacement = expandSourceTriples(displacement, renderToSource);
-
-    const posAttr = geom.getAttribute("position");
-    if (posAttr && posAttr.count === renderToSource.length) {
-        (posAttr.array as Float32Array).set(renderPositions);
-        posAttr.needsUpdate = true;
-    }
-    geom.morphAttributes.position = [new THREE.BufferAttribute(renderDisplacement, 3)];
-    geom.morphTargetsRelative = true;
-
-    // Share the main mesh's influences array so a single write to
-    // mesh.morphTargetInfluences[0] (manual drag or RAF sweep)
-    // moves both meshes. Same trick the line wireframe overlay uses.
-    if (main.morphTargetInfluences) {
-        beamSolid.morphTargetInfluences = main.morphTargetInfluences;
-        beamSolid.morphTargetDictionary = main.morphTargetDictionary ?? undefined;
-    } else if (!beamSolid.morphTargetInfluences) {
-        beamSolid.morphTargetInfluences = [0];
-        beamSolid.morphTargetDictionary = {displacement: 0};
-    }
-
-    // The beam-solid element-edge wireframe: same story as the main mesh's, and
-    // the one the user sees as black lines hanging in space. Its index is written
-    // against the ORIGINAL beam-solid vertices and it holds the position attribute
-    // this geometry had before the element-local expansion swapped one in, so the
-    // unexpanded `displacement` is what fits it. Installed BEFORE
-    // linkLineMorphToMesh runs, which then leaves it alone precisely because it
-    // brought its own.
-
-    // Enable morph targets on every material slot so the GPU
-    // actually applies the delta. The PBR material from the GLB
-    // defaults to morphTargets=false.
-    const enableMorph = (mat: THREE.Material) => {
-        if ("morphTargets" in mat && (mat as unknown as {morphTargets: unknown}).morphTargets !== true) {
-            (mat as unknown as {morphTargets: boolean}).morphTargets = true;
-            mat.needsUpdate = true;
-        }
-    };
-    if (Array.isArray(beamSolid.material)) beamSolid.material.forEach(enableMorph);
-    else if (beamSolid.material) enableMorph(beamSolid.material as THREE.Material);
-
-    // Same dispose dance as applyField: drop the cached morph texture
-    // so three.js rebuilds it from the fresh BufferAttribute on the
-    // next render.
-    geom.dispatchEvent({type: "dispose"});
-}
-
-function snapshotBasePositions(geometry: THREE.BufferGeometry): Float32Array {
-    const attr = geometry.getAttribute("position");
-    if (!attr || attr.itemSize !== 3) {
-        throw new Error("FEA mesh GLB has no usable position attribute");
-    }
-    return new Float32Array(attr.array as Float32Array);
-}
 
 /** Load the mesh GLB, fetch the chosen field's blob, and apply the
  * (component, step) selection. Subsequent calls for the same source
@@ -779,145 +378,15 @@ export async function load_fea_streaming(args: {
             // sections switched off. The bake still writes beam_solids_edges_url;
             // it is simply no longer consumed.
 
-            // AFBV: per-vertex (node0, node1, t). Required so the
-
-            // solid mesh deforms with the rest of the structure
-            // when warp is applied. Best-effort fetch — without it
-            // the solid mesh still renders but stays at base
-            // positions under any morph scale (the old behaviour).
-            if (manifest.mesh.beam_solids_warp_url) {
-                try {
-                    const warp = await fetchBeamSolidsWarp(
-                        fetcher, manifest.mesh.beam_solids_warp_url,
-                    );
-                    if (warp.n_verts === beamSolid.basePositions.length / 3) {
-                        session.active.beamSolidWarp = warp;
-                    } else {
-                        // eslint-disable-next-line no-console
-                        console.warn(
-                            `[fea-streaming] AFBV vertex count ${warp.n_verts} ` +
-                            `!= solid mesh vertices ${beamSolid.basePositions.length / 3}; ` +
-                            `solid beams won't follow deformation.`,
-                        );
-                    }
-                } catch (err) {
-                    // eslint-disable-next-line no-console
-                    console.warn("[fea-streaming] failed to load AFBV warp sidecar:", err);
-                }
-            }
+            const warp = await fetchBeamSolidWarpSidecar(fetcher, manifest, beamSolid.basePositions);
+            if (warp) session.active.beamSolidWarp = warp;
         }
 
-        // Element-edge wireframe overlay. The bake emits an explicit
-        // edge sidecar (deduped uint32 pairs from each cell's
-        // ElemShape.edges) so the wireframe shows real element
-        // boundaries — not the diagonals from quad-face triangulation.
-        // Sharing the mesh's position attribute + morph attribute +
-        // influences array means deformation drives both face and
-        // line rendering from a single buffer / single uniform.
-        if (manifest.mesh.edges_url && !usePerfStore.getState().hideElementEdges) {
-            try {
-                const edgeIndices = await fetchMeshEdges(
-                    fetcher,
-                    manifest.mesh.edges_url,
-                );
-                if (session.active) session.active.edgeIndices = edgeIndices;
-
-                // Beams get their own, dimmer colour. A shell's element edges are a
-                // grid you read element size off; a beam's edge is a member. In one
-                // colour the members vanish into the grid, so the bake now says
-                // which edges are which and they are drawn as two overlays.
-                //
-                // Split rather than overdrawn: the same pair painted twice at the
-                // same depth is a z-fight, and which colour wins is then decided by
-                // the driver.
-                let lineEdgeIndices: Uint32Array | null = null;
-                if (manifest.mesh.line_edges_url) {
-                    try {
-                        const fetched = await fetchMeshEdges(
-                            fetcher,
-                            manifest.mesh.line_edges_url,
-                        );
-                        if (fetched.length > 0) lineEdgeIndices = fetched;
-                    } catch (err) {
-                        // A missing or unreadable split is not worth failing a load
-                        // over — everything simply stays one colour, as before.
-                        // eslint-disable-next-line no-console
-                        console.warn("[fea-streaming] failed to load line edges:", err);
-                    }
-                }
-                const shellEdgeIndices = lineEdgeIndices
-                    ? withoutEdges(edgeIndices, lineEdgeIndices)
-                    : edgeIndices;
-
-                if (shellEdgeIndices.length > 0) {
-                    const lineGeom = new THREE.BufferGeometry();
-                    lineGeom.setAttribute("position", mesh.geometry.attributes.position);
-                    lineGeom.setIndex(new THREE.BufferAttribute(shellEdgeIndices, 1));
-                    const lineMat = new THREE.LineBasicMaterial({
-                        color: FEA_EDGE_COLOR,
-                        depthTest: true,
-                        // Transparent (opacity stays 1 — colour unchanged) so the
-                        // element-edge wireframe joins the transparent render pass
-                        // and, with the renderOrder below, sorts ABOVE a plugin
-                        // field/utilisation face overlay instead of being painted
-                        // over + z-fighting it (which read as flicker on the
-                        // element edges). Opaque lines would render in the opaque
-                        // pass, before any transparent overlay draws over them.
-                        transparent: true,
-                    });
-                    const segments = new THREE.LineSegments(lineGeom, lineMat);
-                    segments.name = "fea-element-edges";
-                    // Above field/plugin face overlays (renderOrder 2), below the
-                    // selection highlight (renderOrder 8), so element edges stay
-                    // legible through a field overlay without hiding selection.
-                    segments.renderOrder = 3;
-                    // Clip the element-edge wireframe with the model under section planes,
-                    // including planes enabled before this (awaited) edge fetch finished.
-                    clipWithModel(segments);
-                    // Layer 1: rendered (camera enables layers 0+1) but
-                    // not pickable (setupPointerHandler's raycaster
-                    // explicitly disables layer 1). prepareLoadedModel
-                    // does the same to the GLB's own LineSegments, but
-                    // it runs before this block — our streaming wireframe
-                    // is added afterwards, so we have to set the layer
-                    // ourselves. Without it, shell elements (where line
-                    // and triangle are coplanar) let the line win the
-                    // raycaster's distance race; the click resolves to
-                    // a LineSegments with no unique_key and selection
-                    // fails with "No drawRanges found for key: undefined".
-                    segments.layers.set(1);
-                    // Share the mesh's morph attribute + influences
-                    // array so the line wireframe morphs in lockstep
-                    // with the face mesh. We set this *after* the
-                    // first applyFieldToMesh call below seeds the
-                    // morph attribute — see linkLineMorphToMesh.
-                    mesh.add(segments);
-                }
-
-                // The beam edges, same geometry and morph story, own colour.
-                if (lineEdgeIndices && lineEdgeIndices.length > 0) {
-                    const beamGeom = new THREE.BufferGeometry();
-                    beamGeom.setAttribute("position", mesh.geometry.attributes.position);
-                    beamGeom.setIndex(new THREE.BufferAttribute(lineEdgeIndices, 1));
-                    const beamMat = new THREE.LineBasicMaterial({
-                        color: FEA_BEAM_EDGE_COLOR,
-                        depthTest: true,
-                        transparent: true,
-                    });
-                    const beamSegments = new THREE.LineSegments(beamGeom, beamMat);
-                    beamSegments.name = "fea-beam-element-edges";
-                    beamSegments.renderOrder = 3;
-                    clipWithModel(beamSegments);
-                    beamSegments.layers.set(1);
-                    mesh.add(beamSegments);
-                }
-            } catch (err) {
-                // Wireframe overlay is decorative — log and continue
-                // so a missing/corrupt sidecar doesn't block rendering.
-                // eslint-disable-next-line no-console
-                console.warn("[fea-streaming] failed to load mesh edges:", err);
-            }
-        }
+        // Element-edge wireframe overlays, from the bake's edge sidecar. The
+        // index is kept on the session so the undeformed reference wireframe can
+        // be rebuilt without re-fetching.
+        const edgeIndices = await installElementEdges(mesh, fetcher, manifest);
+        if (edgeIndices && session.active) session.active.edgeIndices = edgeIndices;
     }
 
     stage("loading field data", 0.55);
@@ -1501,53 +970,3 @@ export async function load_fea_with_defaults(sourceName: string): Promise<void> 
     }
 }
 
-/** Wire the LineSegments wireframe child to share morph attributes
- * + influences with the parent mesh, so changing
- * mesh.morphTargetInfluences[0] morphs both. */
-function linkLineMorphToMesh(mesh: THREE.Mesh): void {
-    for (const child of mesh.children) {
-        if (!(child instanceof THREE.LineSegments)) continue;
-        const lineGeom = child.geometry as THREE.BufferGeometry;
-        // A child that brought its own morph keeps it. The element-edge overlay
-        // SHARES the parent's position buffer, so the parent's per-vertex deltas
-        // are exactly what it needs. The result-line renderer does not: it has
-        // two vertices per beam in its own buffer and its own deltas to match.
-        // Forcing the parent's array onto it hands it deltas of the wrong length
-        // read against the wrong vertices, which is why the coloured beams drifted
-        // away from the black outline instead of moving with it.
-        if (
-            lineGeom.morphAttributes.position
-            && lineGeom.morphAttributes.position !== mesh.geometry.morphAttributes.position
-        ) continue;
-        // morphAttributes is per-geometry; sharing the same array of
-        // BufferAttributes makes both geometries reference the same
-        // morph delta data on the GPU.
-        if (mesh.geometry.morphAttributes.position) {
-            lineGeom.morphAttributes.position = mesh.geometry.morphAttributes.position;
-            lineGeom.morphTargetsRelative = mesh.geometry.morphTargetsRelative;
-        }
-        // morphTargetInfluences is per-Object3D; sharing the same
-        // array reference means writes through mesh.morphTargetInfluences
-        // are visible to the line too.
-        if (mesh.morphTargetInfluences) {
-            child.morphTargetInfluences = mesh.morphTargetInfluences;
-            child.morphTargetDictionary = mesh.morphTargetDictionary ?? undefined;
-        }
-        const mat = child.material as THREE.LineBasicMaterial;
-        if (mat && "morphTargets" in mat) {
-            (mat as any).morphTargets = true;
-            mat.needsUpdate = true;
-        }
-        // Mirror the morph-texture rebuild that applyFieldToMesh does
-        // for the parent mesh. lineGeom shares the parent's position
-        // BufferAttribute, so when applyField dispatched 'dispose' on
-        // mesh.geometry, three.js's WebGLAttributes deleted the GPU
-        // buffer for that shared position. lineGeom's VAO still
-        // references the (now-orphaned) buffer ID, which is why the
-        // wireframe vanishes after a step change. Dispatching dispose
-        // here rebuilds lineGeom's VAO + morph texture against the
-        // freshly-uploaded position buffer on the next render. No-op
-        // on the first call (no renderer state yet).
-        lineGeom.dispatchEvent({type: "dispose"});
-    }
-}
